useragent.ts is stylesheets now, and nothing else
The shorthand expansion, the per-element defaults, the spec-default and
inheritance and initial-value tables were cascade machinery living in
the stylesheet module for load order's sake. They move to the cascade,
private but for the expandShorthands the tests exercise; useragent.ts
keeps the UA document sheet and the widget shadow sheets -- CSS text
in, CSS text out, 193 lines, zero logic.
