+++ src/internal/useragent.ts
@@ -1,1279 +1,14 @@
 /**
- * The user-agent stylesheet and the per-element defaults the cascade resolves
- * against, including the sheets scoped to the form widgets' shadow trees.
+ * The user-agent stylesheets, and nothing else: the UA document sheet every
+ * tree scope receives, and the sheets scoped to the form widgets' shadow
+ * trees. CSS text in, CSS text out -- the cascade resolves it like any
+ * author's, one origin lower.
  *
  * It anchors an invariant: no painter emits a terminal attribute that did not
  * come from a computed style. A selection is inverse video because a rule here
  * declares Highlight/HighlightText, which is CSS's spelling of swapping a
  * cell's colors -- load-bearing, not decorative.
  */
-import {CSS_INITIAL_VALUES} from "./cssproperties.js";
-import {stringWidth} from "./text.js";
-
-const HTML_NAMESPACE = "http://www.w3.org/1999/xhtml";
-
-// ---- Shorthand expansion (the UA table is built on it) ----
-const BORDER_STYLE_KEYWORDS = new Set([
-	"none",
-	"hidden",
-	"dotted",
-	"dashed",
-	"solid",
-	"double",
-	"groove",
-	"ridge",
-	"inset",
-	"outset",
-]);
-const LINE_WIDTH_KEYWORDS = new Set(["thin", "medium", "thick"]);
-const EDGES = ["top", "right", "bottom", "left"] as const;
-/** The two ends of a flow-relative axis, in the order a pair shorthand states them. */
-const AXIS_ENDS = ["start", "end"] as const;
-
-const CORNERS = [
-	"top-left",
-	"top-right",
-	"bottom-right",
-	"bottom-left",
-] as const;
-const LIST_STYLE_POSITIONS = new Set(["inside", "outside"]);
-
-/**
- * CSS 1-4 value expansion: [all], [v h], [t h b], [t r b l]. The four corners
- * fill by the same rule, running top-left, top-right, bottom-right,
- * bottom-left.
- */
-function perEdge(values: string[]): [string, string, string, string] {
-	const [a, b = a, c = a, d = b] = values;
-	return [a, b, c, d];
-}
-
-/** A flow-relative pair's two values: [both] or [start end]. */
-function perEnd(values: string[]): [string, string] {
-	const [a, b = a] = values;
-	return [a, b];
-}
-
-/**
- * A value's top-level components. Whitespace INSIDE parentheses separates a
- * function's own arguments, not components: `rgb(95, 175, 255)` is one color,
- * however many spaces it carries.
- */
-function splitComponents(value: string): string[] {
-	const components: string[] = [];
-	let depth = 0;
-	let start = 0;
-	for (let i = 0; i <= value.length; i++) {
-		const char = value[i];
-		if (char === "(") {
-			depth++;
-		} else if (char === ")") {
-			depth--;
-		} else if ((i === value.length || /\s/.test(char)) && depth === 0) {
-			const component = value.slice(start, i).trim();
-			if (component) {
-				components.push(component);
-			}
-			start = i + 1;
-		}
-	}
-	return components;
-}
-
-/**
- * The `<line-width> || <line-style> || <color>` grammar shared by `border`,
- * the per-side border shorthands and `outline`. Components may appear in any
- * order and any may be omitted.
- */
-function splitLineValue(value: string): {
-	width: string | null;
-	lineStyle: string | null;
-	color: string | null;
-} {
-	let width: string | null = null;
-	let lineStyle: string | null = null;
-	let color: string | null = null;
-	for (const token of splitComponents(value)) {
-		if (BORDER_STYLE_KEYWORDS.has(token)) {
-			lineStyle = token;
-		} else if (/^[\d.]/.test(token) || LINE_WIDTH_KEYWORDS.has(token)) {
-			width = token;
-		} else if (token) {
-			color = token;
-		}
-	}
-	return {width, lineStyle, color};
-}
-
-/** The values `flex-direction` takes, which is how `flex-flow` knows one. */
-const FLEX_DIRECTIONS = new Set([
-	"row",
-	"row-reverse",
-	"column",
-	"column-reverse",
-]);
-
-/** The values `flex-wrap` takes. */
-const FLEX_WRAPS = new Set(["nowrap", "wrap", "wrap-reverse"]);
-
-/**
- * The keywords that qualify the alignment keyword after them rather than
- * standing as a value of their own: `safe center`, `first baseline`.
- */
-const ALIGNMENT_QUALIFIERS = new Set(["safe", "unsafe", "first", "last"]);
-
-/**
- * Expand `flex-flow` (css-flexbox-1 §7.1): a direction and a wrap in either
- * order, either one omitted and left at its initial value.
- */
-function expandFlexFlow(value: string): Record<string, string> {
-	const out: Record<string, string> = {
-		"flex-direction": "row",
-		"flex-wrap": "nowrap",
-	};
-	for (const token of splitComponents(value)) {
-		const keyword = token.toLowerCase();
-		if (FLEX_DIRECTIONS.has(keyword)) {
-			out["flex-direction"] = keyword;
-		} else if (FLEX_WRAPS.has(keyword)) {
-			out["flex-wrap"] = keyword;
-		}
-	}
-	return out;
-}
-
-/**
- * Expand a `place-*` shorthand (css-align-3 §10): the block axis first, then
- * the inline axis, and one value stated for both when only one is written.
- * A value is one keyword, or two where the first only qualifies the second.
- */
-function expandPlace(
-	value: string,
-	block: string,
-	inline: string,
-): Record<string, string> {
-	const values: string[] = [];
-	for (const token of splitComponents(value)) {
-		const previous = values[values.length - 1];
-		if (
-			previous !== undefined &&
-			ALIGNMENT_QUALIFIERS.has(previous.toLowerCase())
-		) {
-			values[values.length - 1] = `${previous} ${token}`;
-		} else {
-			values.push(token);
-		}
-	}
-	if (values.length === 0) {
-		return {};
-	}
-	return {[block]: values[0], [inline]: values[1] ?? values[0]};
-}
-
-/**
- * Expand the `flex` shorthand (css-flexbox-1 §7.1.1): `none` is 0 0 auto,
- * `auto` 1 1 auto, `initial` 0 1 auto; otherwise the first number is grow,
- * a second number is shrink, anything else is the basis -- and a one-value
- * numeric form (`flex: 1`) sets the basis to 0%, which is what makes it the
- * everyday grow-to-fill declaration.
- */
-function expandFlex(value: string): Record<string, string> | null {
-	const v = value.trim();
-	if (v === "none") {
-		return {"flex-grow": "0", "flex-shrink": "0", "flex-basis": "auto"};
-	}
-	if (v === "auto") {
-		return {"flex-grow": "1", "flex-shrink": "1", "flex-basis": "auto"};
-	}
-	if (v === "initial") {
-		return {"flex-grow": "0", "flex-shrink": "1", "flex-basis": "auto"};
-	}
-	let grow: string | undefined;
-	let shrink: string | undefined;
-	let basis: string | undefined;
-	for (const token of v.split(/\s+/)) {
-		if (/^[\d.]+$/.test(token)) {
-			if (grow === undefined) {
-				grow = token;
-			} else if (shrink === undefined) {
-				shrink = token;
-			} else {
-				return null;
-			}
-		} else if (basis === undefined) {
-			basis = token;
-		} else {
-			return null;
-		}
-	}
-	if (grow === undefined && basis === undefined) {
-		return null;
-	}
-	return {
-		"flex-grow": grow ?? "1",
-		"flex-shrink": shrink ?? "1",
-		"flex-basis": basis ?? (grow !== undefined ? "0%" : "auto"),
-	};
-}
-
-/**
- * Expand the `list-style` shorthand, whose components may appear in any order.
- *
- * `none` is ambiguous -- it sets whichever of type/image has not been given --
- * but for a terminal there are no images, so it always means "no marker".
- */
-function expandListStyle(value: string): Record<string, string> {
-	const parts: Record<string, string> = {};
-	for (const token of value.trim().split(/\s+/)) {
-		if (!token) {
-			continue;
-		}
-		if (LIST_STYLE_POSITIONS.has(token)) {
-			parts["list-style-position"] = token;
-		} else if (token.startsWith("url(")) {
-			parts["list-style-image"] = token;
-		} else {
-			parts["list-style-type"] = token;
-		}
-	}
-	return parts;
-}
-
-/**
- * Expand the `background` shorthand down to the two components a terminal can
- * render. Positions, repeats and attachments mean nothing on a cell grid and
- * are dropped; `none` is the IMAGE component, never a color, so the color a
- * bare `background: none` declares is its initial, transparent.
- */
-function expandBackground(value: string): Record<string, string> {
-	const tokens = splitComponents(value);
-	if (value.includes("url(")) {
-		return {"background-image": value.trim()};
-	}
-	const color = tokens
-		.filter((token) => token.toLowerCase() !== "none")
-		.join(" ");
-	return {
-		"background-image": "none",
-		"background-color": color || "transparent",
-	};
-}
-
-/** The four ways a border image tiles, which name the repeat component. */
-const BORDER_IMAGE_REPEATS = new Set(["stretch", "repeat", "round", "space"]);
-
-/** An image value: a function that produces one, or the keyword for none. */
-const IMAGE_VALUE =
-	/^(?:none$|(?:url|(?:repeating-)?(?:linear|radial|conic)-gradient|image|image-set|element|cross-fade|paint)\()/i;
-
-/**
- * Expand the `border-image` shorthand, whose slash-separated groups are the
- * slice, the width and the outset, and whose first group holds the source,
- * the slice and the repeat in any order.
- *
- * A terminal draws no border image, so nothing here reaches the painter. The
- * `border` shorthand resets these five longhands and serializes only while
- * they stand at their initial values, so a declaration block has to know what
- * they hold.
- */
-function expandBorderImage(value: string): Record<string, string> {
-	const out: Record<string, string> = {};
-	const groups = value.split("/").map((group) => group.trim());
-	const slice: string[] = [];
-	const repeat: string[] = [];
-	for (const token of splitComponents(groups[0] ?? "")) {
-		if (BORDER_IMAGE_REPEATS.has(token.toLowerCase())) {
-			repeat.push(token);
-		} else if (IMAGE_VALUE.test(token)) {
-			out["border-image-source"] = token;
-		} else {
-			slice.push(token);
-		}
-	}
-	if (slice.length > 0) {
-		out["border-image-slice"] = slice.join(" ");
-	}
-	if (repeat.length > 0) {
-		out["border-image-repeat"] = repeat.join(" ");
-	}
-	if (groups[1]) {
-		out["border-image-width"] = groups[1];
-	}
-	if (groups[2]) {
-		out["border-image-outset"] = groups[2];
-	}
-	return out;
-}
-
-/**
- * Expand the `border-radius` shorthand: up to four horizontal radii and,
- * after a slash, up to four vertical ones, each list filled out by the CSS
- * 1-4 rule. A corner is elliptical, so its longhand holds both radii -- and
- * states one value where the two agree, which is how a radius serializes.
- */
-function expandBorderRadius(value: string): Record<string, string> {
-	const [across, down] = value.split("/");
-	const horizontal = splitComponents(across ?? "");
-	if (horizontal.length === 0) {
-		return {};
-	}
-	const vertical = down === undefined ? horizontal : splitComponents(down);
-	const horizontalCorners = perEdge(horizontal);
-	const verticalCorners = perEdge(
-		vertical.length === 0 ? horizontal : vertical,
-	);
-	const out: Record<string, string> = {};
-	CORNERS.forEach((corner, i) => {
-		const h = horizontalCorners[i];
-		const v = verticalCorners[i];
-		out[`border-${corner}-radius`] = h === v ? h : `${h} ${v}`;
-	});
-	return out;
-}
-
-/**
- * A value's top-level components, split on `/` rather than on whitespace, with
- * strings, brackets and parentheses kept whole. The grid shorthands are the
- * only ones whose parts are slash-separated and whose components can contain
- * a slash-free quoted string.
- */
-function splitSlashGroups(value: string): string[] {
-	const groups: string[] = [];
-	let depth = 0;
-	let quote = "";
-	let start = 0;
-	for (let i = 0; i < value.length; i++) {
-		const char = value[i];
-		if (quote) {
-			if (char === quote) {
-				quote = "";
-			}
-			continue;
-		}
-		if (char === '"' || char === "'") {
-			quote = char;
-		} else if (char === "(" || char === "[") {
-			depth++;
-		} else if (char === ")" || char === "]") {
-			depth--;
-		} else if (char === "/" && depth === 0) {
-			groups.push(value.slice(start, i).trim());
-			start = i + 1;
-		}
-	}
-	groups.push(value.slice(start).trim());
-	return groups;
-}
-
-/** A value's top-level components, with strings and bracketed names kept whole. */
-function splitGridComponents(value: string): string[] {
-	const components: string[] = [];
-	let depth = 0;
-	let quote = "";
-	let start = 0;
-	for (let i = 0; i <= value.length; i++) {
-		const char = value[i];
-		if (quote) {
-			if (char === quote) {
-				quote = "";
-			}
-			continue;
-		}
-		if (char === '"' || char === "'") {
-			quote = char;
-			continue;
-		}
-		if (char === "(" || char === "[") {
-			depth++;
-		} else if (char === ")" || char === "]") {
-			depth--;
-		} else if ((i === value.length || /\s/.test(char)) && depth === 0) {
-			const component = value.slice(start, i).trim();
-			if (component) {
-				components.push(component);
-			}
-			start = i + 1;
-		}
-	}
-	return components;
-}
-
-/** Whether a grid-placement component is a `<custom-ident>` and nothing else. */
-function isCustomIdent(value: string): boolean {
-	return (
-		/^-?[A-Za-z_][\w-]*$/.test(value) && value !== "auto" && value !== "span"
-	);
-}
-
-/**
- * `grid-row` / `grid-column` (css-grid-2 §8.3.2): a start line and an end
- * line. An omitted end repeats the start only when the start is a name --
- * `grid-column: main` means the whole area called main, while
- * `grid-column: 2` means one track starting at line 2.
- */
-function expandGridPlacementPair(
-	value: string,
-	start: string,
-	end: string,
-): Record<string, string> {
-	const groups = splitSlashGroups(value);
-	const first = groups[0] || "auto";
-	const second =
-		groups.length > 1 && groups[1] ?
-			groups[1] :
-			isCustomIdent(first) ?
-				first :
-				"auto";
-	return {[start]: first, [end]: second};
-}
-
-/**
- * `grid-area` (css-grid-2 §8.4): row-start / column-start / row-end /
- * column-end, each omitted value falling back to the one across from it by the
- * same custom-ident rule the pair shorthands use.
- */
-function expandGridArea(value: string): Record<string, string> {
-	const groups = splitSlashGroups(value);
-	const rowStart = groups[0] || "auto";
-	const fallback = (index: number, from: string): string =>
-		groups.length > index && groups[index] ?
-			groups[index] :
-			isCustomIdent(from) ?
-				from :
-				"auto";
-	const columnStart = fallback(1, rowStart);
-	const rowEnd = fallback(2, rowStart);
-	const columnEnd = fallback(3, columnStart);
-	return {
-		"grid-row-start": rowStart,
-		"grid-column-start": columnStart,
-		"grid-row-end": rowEnd,
-		"grid-column-end": columnEnd,
-	};
-}
-
-/**
- * `grid-template` (css-grid-2 §7.4). Two forms: rows `/` columns, and the
- * visual one whose rows are written as strings of area names with each row's
- * track size and line names around them.
- */
-function expandGridTemplate(value: string): Record<string, string> {
-	const text = value.trim();
-	if (!text || text === "none") {
-		return {
-			"grid-template-rows": "none",
-			"grid-template-columns": "none",
-			"grid-template-areas": "none",
-		};
-	}
-
-	if (!text.includes('"') && !text.includes("'")) {
-		const groups = splitSlashGroups(text);
-		return {
-			"grid-template-rows": groups[0] || "none",
-			"grid-template-columns": groups[1] || "none",
-			"grid-template-areas": "none",
-		};
-	}
-
-	// The visual form: everything up to the last top-level slash states the
-	// rows, and the slash group after it -- if any -- states the columns.
-	const groups = splitSlashGroups(text);
-	const rowsText = groups[0];
-	const columns = groups.length > 1 ? groups.slice(1).join(" / ") : "none";
-
-	const strings: string[] = [];
-	const rowTracks: string[] = [];
-	let pendingNames: string[] = [];
-	let sawString = false;
-	for (const component of splitGridComponents(rowsText)) {
-		if (component.startsWith("[")) {
-			pendingNames.push(component);
-			continue;
-		}
-		if (component.startsWith('"') || component.startsWith("'")) {
-			// A row's own track size follows its string; a row with none is
-			// `auto`, and is written out so the track list stays positional.
-			if (sawString && rowTracks.length < strings.length) {
-				rowTracks.push("auto");
-			}
-			strings.push(component);
-			rowTracks.push(...pendingNames);
-			pendingNames = [];
-			sawString = true;
-			continue;
-		}
-		rowTracks.push(component);
-	}
-	if (sawString && rowTracks.length < strings.length) {
-		rowTracks.push("auto");
-	}
-	rowTracks.push(...pendingNames);
-
-	return {
-		"grid-template-rows": rowTracks.length > 0 ? rowTracks.join(" ") : "none",
-		"grid-template-columns": columns,
-		"grid-template-areas": strings.length > 0 ? strings.join(" ") : "none",
-	};
-}
-
-/**
- * `grid` (css-grid-2 §7.4): the whole explicit grid, or one axis of it
- * against an `auto-flow` that sizes the other's implicit tracks. Either way
- * the shorthand resets every longhand it stands for, which is what makes it
- * safe to write once at the top of a rule.
- */
-function expandGrid(value: string): Record<string, string> {
-	const text = value.trim();
-	const reset = {
-		"grid-auto-flow": "row",
-		"grid-auto-rows": "auto",
-		"grid-auto-columns": "auto",
-	};
-	if (!/\bauto-flow\b/.test(text)) {
-		return {...expandGridTemplate(text), ...reset};
-	}
-
-	const groups = splitSlashGroups(text);
-	if (groups.length !== 2) {
-		return {...expandGridTemplate(text), ...reset};
-	}
-
-	const flowInSecond = /\bauto-flow\b/.test(groups[1]);
-	const flowGroup = flowInSecond ? groups[1] : groups[0];
-	const otherGroup = flowInSecond ? groups[0] : groups[1];
-	const dense = /\bdense\b/.test(flowGroup);
-	const sizes = splitGridComponents(flowGroup)
-		.filter((token) => token !== "auto-flow" && token !== "dense")
-		.join(" ");
-
-	// The axis the flow runs along takes the implicit sizes; the other axis
-	// takes the explicit track list written across the slash.
-	return flowInSecond ?
-			{
-				"grid-template-rows": otherGroup || "none",
-				"grid-template-columns": "none",
-				"grid-template-areas": "none",
-				"grid-auto-flow": dense ? "column dense" : "column",
-				"grid-auto-columns": sizes || "auto",
-				"grid-auto-rows": "auto",
-			} :
-			{
-				"grid-template-columns": otherGroup || "none",
-				"grid-template-rows": "none",
-				"grid-template-areas": "none",
-				"grid-auto-flow": dense ? "row dense" : "row",
-				"grid-auto-rows": sizes || "auto",
-				"grid-auto-columns": "auto",
-			};
-}
-
-/**
- * Expand CSS SHORTHANDS into the longhands everything downstream reads.
- * Declarations are consulted per-property, so a `border: 1px solid` that
- * never becomes border-top-width etc. simply doesn't exist to the box model
- * or the border painter. Declaration order is preserved, so an explicit
- * longhand after a shorthand still overrides it.
- *
- * A shorthand keeps its own entry alongside the longhands it declares, so
- * `getPropertyValue("border")` still answers what was authored. `margin`,
- * `padding` and `border-radius` are the exception: their computed answers are
- * serialized from the four longhands, so keeping the shorthand would shadow
- * that.
- */
-export function expandShorthands(
-	declarations: Record<string, string>,
-): Record<string, string> {
-	const out: Record<string, string> = {};
-	const setEdges = (kind: string, values: string[]) => {
-		const edgeValues = perEdge(values);
-		EDGES.forEach((edge, i) => {
-			out[`border-${edge}-${kind}`] = edgeValues[i];
-		});
-	};
-
-	for (const [property, value] of Object.entries(declarations)) {
-		const values = splitComponents(value);
-		switch (property) {
-			case "border": {
-				const {width, lineStyle, color} = splitLineValue(value);
-				setEdges("width", [width ?? "medium"]);
-				setEdges("style", [lineStyle ?? "none"]);
-				if (color) {
-					setEdges("color", [color]);
-				}
-				break;
-			}
-			case "border-image":
-				Object.assign(out, expandBorderImage(value));
-				break;
-			case "border-width":
-				setEdges("width", values);
-				break;
-			case "border-style":
-				setEdges("style", values);
-				break;
-			case "border-color":
-				setEdges("color", values);
-				break;
-			case "border-radius": {
-				const corners = expandBorderRadius(value);
-				if (Object.keys(corners).length === 0) {
-					break;
-				}
-				Object.assign(out, corners);
-				// The shorthand itself is serialized from these on read.
-				continue;
-			}
-			case "border-top":
-			case "border-right":
-			case "border-bottom":
-			case "border-left": {
-				const edge = property.slice("border-".length);
-				const {width, lineStyle, color} = splitLineValue(value);
-				out[`border-${edge}-width`] = width ?? "medium";
-				out[`border-${edge}-style`] = lineStyle ?? "none";
-				if (color) {
-					out[`border-${edge}-color`] = color;
-				}
-				break;
-			}
-			case "outline": {
-				const {width, lineStyle, color} = splitLineValue(value);
-				out["outline-width"] = width ?? "medium";
-				out["outline-style"] = lineStyle ?? "none";
-				if (color) {
-					out["outline-color"] = color;
-				}
-				break;
-			}
-			// The flow-relative pairs (css-logical-1 §4): one value for both
-			// ends of the axis, or one for each. They state their longhands
-			// and, like `margin` and `padding`, are serialized back from them.
-			case "margin-block":
-			case "margin-inline":
-			case "padding-block":
-			case "padding-inline":
-			case "inset-block":
-			case "inset-inline": {
-				const axis = property.slice(property.lastIndexOf("-") + 1);
-				const kind = property.slice(0, property.lastIndexOf("-"));
-				const ends = perEnd(values);
-				AXIS_ENDS.forEach((end, i) => {
-					out[`${kind}-${axis}-${end}`] = ends[i];
-				});
-				continue;
-			}
-			// `border-block` / `border-inline`: one line drawn on both ends of
-			// the axis.
-			case "border-block":
-			case "border-inline": {
-				const axis = property.slice("border-".length);
-				const {width, lineStyle, color} = splitLineValue(value);
-				for (const end of AXIS_ENDS) {
-					out[`border-${axis}-${end}-width`] = width ?? "medium";
-					out[`border-${axis}-${end}-style`] = lineStyle ?? "none";
-					if (color) {
-						out[`border-${axis}-${end}-color`] = color;
-					}
-				}
-				break;
-			}
-			// One flow-relative edge's line: `border-inline-start: 1px solid`.
-			case "border-block-start":
-			case "border-block-end":
-			case "border-inline-start":
-			case "border-inline-end": {
-				const edge = property.slice("border-".length);
-				const {width, lineStyle, color} = splitLineValue(value);
-				out[`border-${edge}-width`] = width ?? "medium";
-				out[`border-${edge}-style`] = lineStyle ?? "none";
-				if (color) {
-					out[`border-${edge}-color`] = color;
-				}
-				break;
-			}
-			// One line component across an axis: `border-inline-width: 1px 2px`.
-			case "border-block-width":
-			case "border-block-style":
-			case "border-block-color":
-			case "border-inline-width":
-			case "border-inline-style":
-			case "border-inline-color": {
-				const [, axis, kind] = property.split("-");
-				const ends = perEnd(values);
-				AXIS_ENDS.forEach((end, i) => {
-					out[`border-${axis}-${end}-${kind}`] = ends[i];
-				});
-				break;
-			}
-			case "padding":
-			case "margin": {
-				const edgeValues = perEdge(values);
-				EDGES.forEach((edge, i) => {
-					out[`${property}-${edge}`] = edgeValues[i];
-				});
-				// The shorthand itself is serialized from these on read.
-				continue;
-			}
-			case "inset": {
-				const edgeValues = perEdge(values);
-				EDGES.forEach((edge, i) => {
-					out[edge] = edgeValues[i];
-				});
-				break;
-			}
-			case "gap": {
-				out["row-gap"] = values[0];
-				out["column-gap"] = values[1] ?? values[0];
-				break;
-			}
-			// The legacy spelling of `gap`, which browsers still accept. Its
-			// longhands share a cascade slot with row-gap/column-gap, so
-			// declaring them is declaring the modern pair.
-			case "grid-gap": {
-				out["grid-row-gap"] = values[0];
-				out["grid-column-gap"] = values[1] ?? values[0];
-				break;
-			}
-			case "grid-row":
-				Object.assign(
-					out,
-					expandGridPlacementPair(value, "grid-row-start", "grid-row-end"),
-				);
-				break;
-			case "grid-column":
-				Object.assign(
-					out,
-					expandGridPlacementPair(
-						value,
-						"grid-column-start",
-						"grid-column-end",
-					),
-				);
-				break;
-			case "grid-area":
-				Object.assign(out, expandGridArea(value));
-				break;
-			case "grid-template":
-				Object.assign(out, expandGridTemplate(value));
-				break;
-			case "grid":
-				Object.assign(out, expandGrid(value));
-				break;
-			case "overflow": {
-				out["overflow-x"] = values[0];
-				out["overflow-y"] = values[1] ?? values[0];
-				break;
-			}
-			case "flex": {
-				Object.assign(out, expandFlex(value) ?? {});
-				break;
-			}
-			case "flex-flow": {
-				Object.assign(out, expandFlexFlow(value));
-				break;
-			}
-			case "place-content":
-				Object.assign(
-					out,
-					expandPlace(value, "align-content", "justify-content"),
-				);
-				break;
-			case "place-items":
-				Object.assign(out, expandPlace(value, "align-items", "justify-items"));
-				break;
-			case "place-self":
-				Object.assign(out, expandPlace(value, "align-self", "justify-self"));
-				break;
-			case "list-style":
-				Object.assign(out, expandListStyle(value));
-				break;
-			case "background":
-				Object.assign(out, expandBackground(value));
-				break;
-			case "text-decoration": {
-				// `<line> || <style> || <color>`; only the line component has a
-				// terminal rendering, and it is the one the painter reads.
-				const line = values
-					.filter((token) =>
-						/^(none|underline|overline|line-through|blink)$/.test(token),
-					)
-					.join(" ");
-				if (line) {
-					out["text-decoration-line"] = line;
-				}
-				break;
-			}
-		}
-		out[property] = value;
-	}
-	return out;
-}
-
-// ---- CSS spec defaults ----
-/**
- * CSS specification defaults for properties
- */
-const CSS_SPEC_DEFAULTS: Record<string, string> = {
-	"display": "inline",
-	"margin-top": "0",
-	"margin-right": "0",
-	"margin-bottom": "0",
-	"margin-left": "0",
-	"padding-top": "0",
-	"padding-right": "0",
-	"padding-bottom": "0",
-	"padding-left": "0",
-	"border-width": "0",
-	"border-style": "none",
-	"border-color": "currentColor",
-	"border-top-width": "0",
-	"border-right-width": "0",
-	"border-bottom-width": "0",
-	"border-left-width": "0",
-	"border-top-style": "none",
-	"border-right-style": "none",
-	"border-bottom-style": "none",
-	"border-left-style": "none",
-	"border-top-color": "currentColor",
-	"border-right-color": "currentColor",
-	"border-bottom-color": "currentColor",
-	"border-left-color": "currentColor",
-	"background-color": "transparent",
-	"color": "#000000",
-	// One cell tall: the terminal's font is the grid, and a length written in
-	// em is a length written in cells.
-	"font-size": "1px",
-	"font-weight": "normal",
-	"font-style": "normal",
-	"text-decoration": "none",
-	"white-space": "normal",
-	"overflow": "visible",
-	"position": "static",
-	"width": "auto",
-	"height": "auto",
-	"box-sizing": "border-box",
-	// Terminal-optimized flexbox defaults
-	// Container properties
-	"flex-direction": "row",
-	"flex-wrap": "nowrap",
-	// `normal` is CSS's own initial value on both, and the one grid needs: it
-	// is what tells a grid's auto tracks to fill the container. In a flex
-	// container it means what flex-start meant here before, so nothing about
-	// flex layout moves.
-	"justify-content": "normal",
-	"align-items": "stretch",
-	"align-content": "normal",
-	"gap": "0",
-	"row-gap": "0",
-	"column-gap": "0",
-	// Item properties
-	"flex-grow": "0",
-	"flex-shrink": "1",
-	"flex-basis": "auto",
-	"align-self": "auto",
-	"order": "0",
-};
-
-// ---- Terminal element defaults ----
-const TERMINAL_ELEMENT_DEFAULTS: Record<string, Record<string, string>> = {
-	// Metadata elements - never rendered in terminal
-	head: {display: "none"},
-	link: {display: "none"},
-	meta: {display: "none"},
-	script: {display: "none"},
-	style: {display: "none"},
-	title: {display: "none"},
-
-	// Block elements
-	article: {display: "block"},
-	aside: {display: "block"},
-	blockquote: {display: "block"},
-	body: {display: "block"},
-	dd: {display: "block"},
-	// A disclosure and its summary are both blocks, so the summary owns its
-	// row and the body stacks under it. The marker and the open/closed glyph
-	// are ::before rules in the UA document stylesheet.
-	details: {display: "block"},
-	summary: {display: "block", cursor: "pointer"},
-	div: {display: "block"},
-	dl: {display: "block"},
-	dt: {display: "block"},
-	// A group of controls in a labelled box, the way a browser draws one: a
-	// border around the group with the legend sitting IN the top border line.
-	// The legend gets there by rising one row onto the border and painting the
-	// terminal's own background over the cells it covers -- which is what
-	// "interrupting the border" is.
-	fieldset: {
-		display: "block",
-		border: "1px solid",
-		padding: "0 1ch",
-	},
-	legend: {
-		"display": "block",
-		"margin-top": "-1px",
-		"font-weight": "bold",
-	},
-	figcaption: {display: "block"},
-	figure: {display: "block"},
-	footer: {display: "block"},
-	form: {display: "block"},
-	h1: {display: "block"},
-	h2: {display: "block"},
-	h3: {display: "block"},
-	h4: {display: "block"},
-	h5: {display: "block"},
-	h6: {display: "block"},
-	header: {display: "block"},
-	hr: {"display": "block", "border-top": "1px solid"},
-	html: {display: "block"},
-	li: {display: "list-item"},
-	main: {display: "block"},
-	nav: {display: "block"},
-	ol: {"display": "block", "padding-left": "4ch"},
-	p: {display: "block"},
-	pre: {"display": "block", "white-space": "pre"},
-	// A gauge is a flat field in the input family, sized like a browser's own
-	// unstyled progress bar: a fixed track the fill is a fraction of.
-	progress: {"display": "inline-block", "width": "10ch", "white-space": "pre"},
-	meter: {"display": "inline-block", "width": "10ch", "white-space": "pre"},
-	section: {display: "block"},
-	ul: {"display": "block", "padding-left": "4ch"},
-
-	// Inline elements
-	a: {display: "inline"},
-	abbr: {display: "inline"},
-	b: {"display": "inline", "font-weight": "bold"},
-	br: {display: "inline"},
-	cite: {"display": "inline", "font-style": "italic"},
-	code: {"display": "inline", "background-color": "rgba(0, 0, 0, 0.1)"},
-	dfn: {"display": "inline", "font-style": "italic"},
-	em: {"display": "inline", "font-style": "italic"},
-	i: {"display": "inline", "font-style": "italic"},
-	// A key is a keycap: bold text inside brackets, "[q]", the form every
-	// terminal help screen and man page uses for a key to press. A browser
-	// draws the cap with a border and a monospace face; a terminal is already
-	// monospace and cannot afford a box around one glyph, so the accelerator
-	// convention stands in: the marked keys are bold and underlined, the way
-	// CUA-era TUIs marked a menu's mnemonic letter. A decoration costs no
-	// cells, so marking up a key never moves layout.
-	kbd: {
-		"display": "inline",
-		"font-weight": "bold",
-		"text-decoration": "underline",
-	},
-	label: {display: "inline"},
-	mark: {display: "inline"},
-	q: {display: "inline"},
-	s: {"display": "inline", "text-decoration": "line-through"},
-	samp: {display: "inline"},
-	// As in browsers: a slot generates no box of its own -- its projected
-	// (or fallback) content is spliced into the parent's child sequence by
-	// the walker's flat-tree layer (see composition.ts). Styling the slot
-	// still works for inherited properties, exactly the browser behavior.
-	slot: {display: "contents"},
-	// SGR faint is the terminal's small: same glyph cells, reduced ink.
-	small: {"display": "inline", "font-weight": "lighter"},
-	span: {display: "inline"},
-	strong: {"display": "inline", "font-weight": "bold"},
-	sub: {display: "inline"},
-	sup: {display: "inline"},
-	time: {display: "inline"},
-	u: {"display": "inline", "text-decoration": "underline"},
-	var: {"display": "inline", "font-style": "italic"},
-
-	// Terminal UI controls. The button joins the flat field family: no
-	// border (three rows and two columns per button, in a world of one-row
-	// list items), just breathing room and the family's focus underline
-	// (see getElementDefaults). Authors who want chrome add it.
-	// The button is the toggles' visual language extended to labels:
-	// "[ Label ]", the delimited one-row form every terminal tradition
-	// from dialog/whiptail to Midnight Commander to the text-mode
-	// browsers (Lynx, w3m, ELinks render HTML buttons exactly this way)
-	// converged on. The brackets are UA ::before/::after rules in the UA
-	// document stylesheet -- author content rules override them (an
-	// icon button sets its own), and focus underlines the whole token
-	// like the rest of the field family.
-	button: {
-		display: "inline-block",
-		cursor: "pointer",
-	},
-	// A dialog is a box drawn over the page, so it is bordered and opaque:
-	// the border is what says where the dialog ends and the document it sits
-	// on begins, and Canvas -- the terminal's own background -- is what makes
-	// a non-modal one, which has no backdrop clearing the viewport for it,
-	// still hide the content it covers rather than tangle with it.
-	dialog: {
-		"display": "block",
-		"border": "1px solid",
-		"padding": "0 1ch",
-		"background-color": "Canvas",
-	},
-	// A text input is a flat field: bare when blurred (dim placeholder and
-	// the content are the affordance -- the convention of the entire
-	// prompt-tool ecosystem), underlined when FOCUSED (see
-	// getElementDefaults) -- "underline means live." Plain SGR 4 only:
-	// styled underlines (4:2) verified dead through the baseline
-	// tmux+Terminal.app chain, where the intermediary normalizes the
-	// graceful 4-then-4:2 pair into one styled attribute and the terminal
-	// drops it entirely -- the focused field would lose its marker on
-	// exactly the stack we promise works. No borders (three rows and two
-	// columns per field), no backgrounds (no theme-safe color exists).
-	// Width mirrors the browser's own size=20 default. This is the UA
-	// baseline, deliberately lightweight; authors who want chrome add it.
-	input: {
-		"display": "inline-block",
-		"width": "20ch",
-		// A field's value never wraps or collapses -- runs of spaces are
-		// real content, and the painter's scroll-window handles overflow.
-		"white-space": "pre",
-	},
-	// A select is a flat field in the input family: the selected option's
-	// label plus a dim indicator, underlined when focused (see
-	// getElementDefaults for the dynamic width and focus underline).
-	select: {
-		"display": "inline-block",
-		"white-space": "pre",
-	},
-	// A textarea preserves newlines and soft-wraps at its edge, exactly the
-	// browser default. Its UA shadow tree's value text lays out through the
-	// normal pipeline, so this is what makes multiline values multiline --
-	// and break-word (the browser's own textarea UA rule) is what makes a
-	// long unbroken word wrap at the field edge instead of escaping it.
-	textarea: {
-		"display": "inline-block",
-		"border": "1px solid",
-		"padding": "0 1ch",
-		"white-space": "pre-wrap",
-		"overflow-wrap": "break-word",
-	},
-
-	// Tables
-	caption: {display: "table-caption"},
-	col: {display: "table-column"},
-	colgroup: {display: "table-column-group"},
-	table: {"display": "table", "border-collapse": "collapse"},
-	tbody: {display: "table-row-group"},
-	td: {
-		"display": "table-cell",
-		"border-top-width": "1px",
-		"border-right-width": "1px",
-		"border-bottom-width": "1px",
-		"border-left-width": "1px",
-		"border-top-style": "solid",
-		"border-right-style": "solid",
-		"border-bottom-style": "solid",
-		"border-left-style": "solid",
-		"padding-left": "1ch",
-		"padding-right": "1ch",
-	},
-	tfoot: {display: "table-footer-group"},
-	th: {
-		"display": "table-cell",
-		"border-top-width": "1px",
-		"border-right-width": "1px",
-		"border-bottom-width": "1px",
-		"border-left-width": "1px",
-		"border-top-style": "solid",
-		"border-right-style": "solid",
-		"border-bottom-style": "solid",
-		"border-left-style": "solid",
-		"padding-left": "1ch",
-		"padding-right": "1ch",
-		"font-weight": "bold",
-	},
-	thead: {display: "table-header-group"},
-	tr: {display: "table-row"},
-};
-
-// The defaults above may use shorthands; normalize them once so the
-// per-property consultation below always finds longhands.
-for (const [tag, declarations] of Object.entries(TERMINAL_ELEMENT_DEFAULTS)) {
-	TERMINAL_ELEMENT_DEFAULTS[tag] = expandShorthands(declarations);
-}
-// input's own entry in TERMINAL_ELEMENT_DEFAULTS above (bordered box, 20ch
-// wide) is shaped for a text field, whose void-element content has nothing
-// else to size or paint a box from. A checkbox/radio renders as a compact
-// "[ ]"/"[x]" glyph instead -- same reasoning, same
-// problem, opposite answer: 3 cells wide, no border, no padding to pad it
-// out further.
-const CHECKBOX_DEFAULTS: Record<string, string> = {
-	display: "inline-block",
-	width: "3ch",
-};
-/**
- * TERMINAL_ELEMENT_DEFAULTS keyed purely by tag name can't distinguish an
- * <input type="checkbox"> from a text <input> -- both are just "input". This
- * is the one place type has to be checked before falling back to the
- * tag-level defaults.
- */
-export function getElementDefaults(
-	element: Element,
-): Record<string, string> | undefined {
-	// The defaults are keyed by the element's own name, which is what the
-	// table is indexed by and what every branch below asks about. `tagName`
-	// answers the same question in the document's case convention, deriving a
-	// fresh string every time it is read; this is read once, per property, per
-	// element, on the resolution path.
-	// The sheet is HTML's. An element of the same local name in another
-	// namespace -- the <select> a stray `<svg>` puts its children under -- is
-	// not that element and has none of its interface.
-	if (element.namespaceURI !== HTML_NAMESPACE) {
-		return undefined;
-	}
-	const name = element.localName;
-	const document = element.ownerDocument;
-	// The browser's UA :fullscreen treatment: the fullscreen element fills
-	// the viewport. Explicit cells rather than percentages -- the alternate
-	// screen IS the containing geometry, and innerWidth/Height are its size.
-	if (document !== null && document.fullscreenElement === element) {
-		const window = document.defaultView;
-		const base = TERMINAL_ELEMENT_DEFAULTS[name] ?? {};
-		if (window) {
-			return {
-				...base,
-				// The browser's :fullscreen block: fixed at the viewport
-				// origin, viewport-sized, opaque over the document (Canvas =
-				// the terminal's own background, the ::backdrop stand-in).
-				"position": "fixed",
-				"top": "0px",
-				"left": "0px",
-				"width": `${window.innerWidth}ch`,
-				"height": `${window.innerHeight}px`,
-				"background-color": "Canvas",
-			};
-		}
-	}
-	// A textarea's rows/cols are NOT defaults here: they size the CONTENT
-	// box, and only layout knows what border and padding the cascade actually
-	// left on the element to add around it. Baking the UA chrome into a
-	// min-height/width constant left authors unable to unbake it with
-	// `border: none`. See the textarea leaf sizing in layout.ts.
-	if (name === "button") {
-		const merged: Record<string, string> = {
-			...TERMINAL_ELEMENT_DEFAULTS.button,
-		};
-		if (document?.activeElement === element) {
-			merged["text-decoration-line"] = "underline";
-		}
-		return merged;
-	}
-	if (name === "select") {
-		// Sized to the LONGEST option label plus the indicator, exactly as a
-		// browser sizes a closed select -- so the field's width never jumps
-		// as the selection changes.
-		const select = element as HTMLSelectElement;
-		let widest = 0;
-		for (const option of select.options) {
-			widest = Math.max(widest, stringWidth(option.label));
-		}
-		const merged: Record<string, string> = {
-			...TERMINAL_ELEMENT_DEFAULTS.select,
-			width: `${widest + 2}ch`,
-		};
-		if (document?.activeElement === select) {
-			merged["text-decoration-line"] = "underline";
-		}
-		return merged;
-	}
-	if (name === "input") {
-		const input = element as HTMLInputElement;
-		// The FOCUSED field gets the underline -- the UA "this is the live
-		// one" signal, in plain SGR 4, the one underline every terminal and
-		// every intermediary renders. Focus changes invalidate the
-		// computed-style cache (see handleFocusChange), so this is
-		// re-consulted at the right moments.
-		const focused = document?.activeElement === input;
-		if (input.type === "checkbox" || input.type === "radio") {
-			// The compact glyph is bare when blurred; focus underlines it --
-			// same live-wire language as the text field.
-			return focused ?
-					{...CHECKBOX_DEFAULTS, "text-decoration-line": "underline"} :
-				CHECKBOX_DEFAULTS;
-		}
-		// The size attribute drives a text input's default width, one column
-		// per character position, exactly as a browser sizes an unstyled
-		// input from size="...". The static defaults entry carries the spec
-		// default of 20.
-		const size = parseInt(input.getAttribute("size") ?? "", 10);
-		if (Number.isFinite(size) || focused) {
-			const merged = {...TERMINAL_ELEMENT_DEFAULTS.input};
-			if (Number.isFinite(size) && size > 0) {
-				merged.width = `${size}ch`;
-			}
-			if (focused) {
-				merged["text-decoration-line"] = "underline";
-			}
-			return merged;
-		}
-	}
-	return TERMINAL_ELEMENT_DEFAULTS[name];
-}
-
-// ---- Inheritance / initial-value tables ----
-/**
- * Properties that inherit by default
- */
-export const INHERITED_PROPERTIES = new Set([
-	"color",
-	"cursor",
-	"direction",
-	"font-family",
-	"font-size",
-	"font-style",
-	"font-variant",
-	"font-weight",
-	"letter-spacing",
-	"line-height",
-	"list-style",
-	"list-style-image",
-	"list-style-position",
-	"list-style-type",
-	"overflow-wrap",
-	"quotes",
-	"text-align",
-	"text-decoration",
-	"text-decoration-color",
-	"text-decoration-line",
-	"text-decoration-style",
-	"text-decoration-thickness",
-	"text-indent",
-	"text-transform",
-	"visibility",
-	"white-space",
-	"word-break",
-	"word-spacing",
-]);
-
-export const INITIAL_KEYWORDS = new Set([
-	"initial",
-	"revert",
-	"revert-layer",
-	"unset",
-]);
-
-/**
- * A property's initial value on an element, or -- with no element, as for a
- * pseudo-element -- the initial value alone, which no element default has a
- * say in.
- */
-export function getInitialStyle(
-	element: Element | null,
-	property: string,
-): string {
-	// Check element-specific defaults first
-	const elementDefaults = element ? getElementDefaults(element) : null;
-	if (elementDefaults && elementDefaults[property]) {
-		return elementDefaults[property];
-	}
-
-	// Fall back to CSS spec default, and past it to the property index --
-	// every longhand has an initial value, and a property this engine does not
-	// lay out still resolves to one.
-	return CSS_SPEC_DEFAULTS[property] || CSS_INITIAL_VALUES[property] || "";
-}
-
 // ---- UA document stylesheet ----
 /**
  * The UA DOCUMENT stylesheet. Rules here are UA origin (every author rule
